pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"

requires = [ "setuptools", "wheel" ]

[project]
name = "pkg-18604"
version = "0.0.0"

readme = "README.md"
license-files = [ "LICENSE" ]
authors = [
  { name = "developers", email = "developers@tagdots.com" },
]
requires-python = ">=3.12.10"

classifiers = [
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Build Tools",
]

dependencies = [
  "click==8.2.1",
  "click-option-group==0.5.7",
  "gitpython==3.1.44",
  "setuptools==80.9.0",
  "virtualenv==20.31.2",
]

optional-dependencies.test = [
  "build",
  "commitizen",
  "coverage",
  "flake8",
  "isort",
  "pre-commit",
  "pyscan-rs",
  "pytest",
  "pytest-cov",
  "pytest-env",
  "twine",
]

urls.Changelog = "https://www.notavailable.com/changelog"
urls.Documentation = "https://www.notavailable.com/doc"
urls.Homepage = "https://www.notavailable.com"

scripts.pkg-18604 = "pkg_18604.run:main"

[dependency-groups]
test = [
  "coverage[toml]",
]

[tool.setuptools.packages.find]
where = [ "src" ]

[tool.setuptools.dynamic]
version = { attr = "src.pkg_18604.__init__" }

[tool.pyproject-fmt]
column_width = 125
indent = 2
keep_full_version = true

[tool.coverage]
source = [ "." ]

[tool.coverage.run]
branch = true
command_line = "-m pytest -v"
include = [ "src/pkg_18604/*" ]
omit = [ "tests/*" ]

[tool.coverage.report]
fail_under = 80
show_missing = false
include = [ "src/pkg_18604/*" ]

[tool.commitizen]
name = "cz_conventional_commits"
major_version_zero = false
tag_format = "$version"
update_changelog_on_bump = true
version_files = [
  "src/pkg_18604/__init__.py:__version__",
]
version_provider = "pep621"
version_scheme = "pep440"
